Sample the Shell Desktop Window, and Enable Context Aware Color Choices
Adds a new option, enabled by default for Desktop measures, to allow
context-aware color shuffling based on the position of the skin. Skins with
this option enabled will use the Chameleon-selected color that most closely
matches the area under the skin as the primary background color, and if
needed will swap the background and foreground colors.

And also fix cropping requiring a refresh to take new crop values into
effect, because I forgot to commit that separately.